fix/api: Ensure we get the most recent bookmarks locations
last_bookmarks_locations() method returned the correct locations, but not from
the most recent to the oldest one as it was expected. This is due to the
'GROUP BY' command returning the bookmarks in an undefined order.

We ensure we are selecting the most recent bookmarks created by getting the
bookmarks with the greatest id.